fix(settings): the catalog and the nav make room for a page
Adding a settings page grows two registries past their frozen line
counts, and the waiver file's own rule is that it absorbs nothing that
grew. Both are genuinely shorter instead:

- The lucide glyph table moves out of settingsnav.tsx into
  settingsnavicons.ts. It was already a self-contained block with its own
  rationale for living outside the React-free catalog.
- The capability-expression builders move out of settingscatalog.ts into
  settingscapability.ts, re-exported so no caller changes. The table and
  the shorthand it is written in had grown past one screen together, and
  the shorthand is the half that does not change when a page is added.

UnitSecretScope and CatalogContext stay in the catalog: a gate reads that
type's declaration out of that file BY PATH, so where it lives is part of
its contract, not an accident of layout.

The exhaustive specs learn the new page. `custom_field` now names three
settings pages rather than two, and both seeded principal lists reach the
acquisition catalog exactly where they reach the lead vocabulary, since
the two are gated on the same object.

The deal's new fields gain history labels, so a change row says "Deal
brief" rather than a raw column name. `history.field.description` already
existed and is shared with other record types, so the mapping reuses it.
