+++ frontend/src/screens/settingscatalog.test.ts
@@ -104,7 +104,7 @@ describe("the scope each page declares", () => {
     pipelines: "workspace",
     stageautomation: "workspace",
     leads: "workspace",
-    "acquisition-sources": "workspace",
+    acquisition: "workspace",
     fields: "workspace",
     tags: "workspace",
     products: "workspace",
@@ -216,7 +216,7 @@ describe("what each page lets a reader change", () => {
     leads:
       "all(full-seat, any(any(custom_field:update, custom_field:create), custom_field:delete))",
     // No delete arm: a source is retired through its switch, never removed.
-    "acquisition-sources":
+    acquisition:
       "all(full-seat, any(any(custom_field:update, custom_field:create)))",
     fields:
       "all(full-seat, any(any(custom_field:update, custom_field:create)))",
@@ -935,7 +935,7 @@ describe("what the rail carries and what it leaves behind", () => {
       "leads",
       // The acquisition-source catalog, gated on custom_field like the lead
       // vocabulary beside it: a rep reads the channels and changes none.
-      "acquisition-sources",
+      "acquisition",
       "fields",
       "tags",
       "knowledge",
