+++ frontend/src/screens/settingsstories.test.ts
@@ -99,7 +99,7 @@ describe("the settings stories are filed where the product files them", () => {
   // So the count is EXACT and derived from the tree: adding or removing a
   // settings story is a deliberate edit to this number.
   it("reads every settings story, and says how many that is", () => {
-    expect(settingsStories.length).toBe(75);
+    expect(settingsStories.length).toBe(76);
   });
 
   // The filter above drops a file whose title does not resolve. That is the
